Moved validateInput to utils
